Update hd44780_I2Cexp.h
I had to add (byte) to the endTransmission statement to get the library to compile for ESP32 in Arduino 1.8.5

+++ hd44780ioClass/hd44780_I2Cexp.h
@@ -373,7 +373,7 @@ static uint8_t AutoInst;
 	{
 		// check to see if device at specified address is really there
 		Wire.beginTransmission(_addr);
-		if(Wire.endTransmission(1))
+		if(Wire.endTransmission((byte)1))
 			return(hd44780::RV_ENXIO);
 	}
 
@@ -416,7 +416,7 @@ static uint8_t AutoInst;
 		 */
 		Wire.write(5);	// point to IOCON
 		Wire.write(0x20);// disable sequential mode (enables BYTE mode)
-		Wire.endTransmission(1);
+		Wire.endTransmission((byte)1);
 
 		/*
 		 * Now set up output port
@@ -424,7 +424,7 @@ static uint8_t AutoInst;
 		Wire.beginTransmission(_addr);
 		Wire.write((uint8_t)0); // point to IODIR
 		Wire.write((uint8_t)0); // all pins output
-		Wire.endTransmission(1);
+		Wire.endTransmission((byte)1);
 	
 		/*
 		 * point chip to GPIO
@@ -434,7 +434,7 @@ static uint8_t AutoInst;
 		
 	}
 	Wire.write((uint8_t)0);  // Set the entire output port to LOW
-	if( (status = Wire.endTransmission(1)) ) // assignment
+	if( (status = Wire.endTransmission((byte)1)) ) // assignment
 		status = hd44780::RV_EIO;
 
 	return ( status );
@@ -497,14 +497,14 @@ int rval = hd44780::RV_EIO;
 	
 	Wire.beginTransmission(_addr);
 	Wire.write(gpioValue);		// d4-d7 are inputs, RS, r/w high, E LOW
-	if(Wire.endTransmission(1))
+	if(Wire.endTransmission((byte)1))
 		goto returnStatus;
 
 
 	// raise E to read the data.
 	Wire.beginTransmission(_addr);
 	Wire.write(gpioValue | _en); // Raises E 
-	if(Wire.endTransmission(1))
+	if(Wire.endTransmission((byte)1))
 		goto returnStatus;
 
 	// read the expander port to get the upper nibble of the byte
@@ -515,7 +515,7 @@ int rval = hd44780::RV_EIO;
 
 	Wire.beginTransmission(_addr);
 	Wire.write(gpioValue); // lower E after reading nibble
-	if(Wire.endTransmission(1))
+	if(Wire.endTransmission((byte)1))
 		goto returnStatus;
 
 	// map i/o expander port bits into upper nibble of byte
@@ -533,7 +533,7 @@ int rval = hd44780::RV_EIO;
 	
 	Wire.beginTransmission(_addr);
 	Wire.write(gpioValue | _en); // Raise E to read next nibble
-	if(Wire.endTransmission(1))
+	if(Wire.endTransmission((byte)1))
 		goto returnStatus;
 
 	// read the expander port to get the lower nibble of the byte
@@ -549,7 +549,7 @@ int rval = hd44780::RV_EIO;
 
 	Wire.beginTransmission(_addr);
 	Wire.write(gpioValue); // lower E after reading nibble
-	if(Wire.endTransmission(1))
+	if(Wire.endTransmission((byte)1))
 		goto returnStatus;
 
 	// map i/o expander port bits into lower nibble of byte
@@ -572,7 +572,7 @@ returnStatus:
 	// try to put gpio port back to all outputs state with WR signal low for writes
 	Wire.beginTransmission(_addr);
 	Wire.write(_blCurState);		// with E LOW
-	if(Wire.endTransmission(1))
+	if(Wire.endTransmission((byte)1))
 		rval = hd44780::RV_EIO;
 
 	return(rval);
@@ -620,7 +620,7 @@ int iowrite(hd44780::iotype type, uint8_t value)
 	{
 		write4bits( (value & 0x0F), type); // lower nibble, if not 4bit cmd
 	}
-	if(Wire.endTransmission(1))
+	if(Wire.endTransmission((byte)1))
 		return(hd44780::RV_EIO);
 
 	return(hd44780::RV_ENOERR);
@@ -651,7 +651,7 @@ int iosetBacklight(uint8_t dimvalue)
 		Wire.write(9); // point to GPIO
 	}
 	Wire.write( _blCurState );
-	if(Wire.endTransmission(1))
+	if(Wire.endTransmission((byte)1))
 		return(hd44780::RV_EIO);
 
 	return(hd44780::RV_ENOERR); // all is good
@@ -712,7 +712,7 @@ uint8_t locinst = 0;
 	for(address = 0x20; address <= 0x27; address++ )
 	{
 		Wire.beginTransmission(address);
-		error = Wire.endTransmission(1);
+		error = Wire.endTransmission((byte)1);
 		// chipkit stuff screws up if you do beginTransmission() too fast
 		// after an endTransmission()
 		// below 20us will cause it to fail
@@ -738,7 +738,7 @@ uint8_t locinst = 0;
 	for(address = 0x38; address <= 0x3f; address++ )
 	{
 		Wire.beginTransmission(address);
-		error = Wire.endTransmission(1);
+		error = Wire.endTransmission((byte)1);
 		// chipkit stuff screws up if you do beginTransmission() too fast
 		// after an endTransmission()
 		// below 20us will cause it to fail.
@@ -807,7 +807,7 @@ I2CexpType chiptype;
 	Wire.beginTransmission(address);
 	Wire.write((uint8_t) 0);	// try to point to MCP23008 IODR
 	Wire.write((uint8_t) 0xff);	// try to write to MCP23008 IODR
-	Wire.endTransmission(1);
+	Wire.endTransmission((byte)1);
 
 	/*
 	 * Now try to point MCP23008 to IODIR for read
@@ -816,7 +816,7 @@ I2CexpType chiptype;
 
 	Wire.beginTransmission(address);
 	Wire.write((uint8_t) 0);	// try to point to MCP23008 IODR
-	Wire.endTransmission(1);
+	Wire.endTransmission((byte)1);
 
 	/*
 	 * Now read a byte
@@ -944,7 +944,7 @@ uint8_t rs, rw, en, d4, d5, d6, d7, bl, blLevel;
 
 	Wire.beginTransmission(_addr);
 	Wire.write((uint8_t) 0xff);
-	Wire.endTransmission(1);
+	Wire.endTransmission((byte)1);
 
 	// now read back from the port
 
@@ -956,7 +956,7 @@ uint8_t rs, rw, en, d4, d5, d6, d7, bl, blLevel;
 	
 	Wire.beginTransmission(_addr);
 	Wire.write((uint8_t) (~(1 << 2)) );
-	Wire.endTransmission(1);
+	Wire.endTransmission((byte)1);
 
 	// read back data
 	Wire.requestFrom((int)_addr, 1);
@@ -994,7 +994,7 @@ uint8_t rs, rw, en, d4, d5, d6, d7, bl, blLevel;
 	
 		Wire.beginTransmission(_addr);
 		Wire.write((uint8_t) (~(1 << 4)) );
-		Wire.endTransmission(1);
+		Wire.endTransmission((byte)1);
 
 		// read back data
 		Wire.requestFrom((int)_addr, 1);
@@ -1094,12 +1094,12 @@ uint8_t blLevel;
 	Wire.beginTransmission(_addr);
 	Wire.write((uint8_t)0); // point to IODIR
 	Wire.write(0xff); // all pins inputs
-	Wire.endTransmission(1);
+	Wire.endTransmission((byte)1);
 
 	Wire.beginTransmission(_addr);
 	Wire.write(6); // point to GPPU
 	Wire.write(0xff); // turn on pullups
-	Wire.endTransmission(1);
+	Wire.endTransmission((byte)1);
 
 	/*
 	 * read from the GPIO port
@@ -1107,7 +1107,7 @@ uint8_t blLevel;
 
 	Wire.beginTransmission(_addr);
 	Wire.write(9); // point to GPIO
-	Wire.endTransmission(1);
+	Wire.endTransmission((byte)1);
 	Wire.requestFrom((int)_addr, 1);
 	data = Wire.read();
